sistema_loja/requirements.txt
mysql-connector-python>=9.0.0
ttkthemes>=3.2.2